explicitly detect iso9660 on a disk (#903158)
Previously we were relying on parted to raise an error when trying to
examine a disk with an iso9660 image written to it. The reason for this
was that isohybrid was writing an incorrect mbr, which has now been
fixed. parted will now return the dos mbr from the hybrid iso.

This skips handleUdevDiskLabel if the device format is iso9660, allowing
it to be selected instead of the mbr.