Rewrite the DeviceFactory classes.
The main change here is the addition of nested factories to enable
more complex stacks like lvm with an md pv and, later, lvm thin
provisioning. This is accomplished by allowing factory classes to specify
a child factory class. For example, MDFactory.child_factory_class is
PartitionSetFactory.

The management of a partition set as is used for lvm, md, and btrfs
member devices has been moved into its own factory class. It is
slightly different from the others in that it cannot have a child
factory and cannot either be a top-level factory.

The management of container devices like lvm vg or btrfs volume is
handled within the factory classes for lvm lv and btrfs subvolume,
respectively. There is not a separate set of factory classes for the
container types since they are closely tied to the leaf device types.

The factories can be used to modify/adjust non-existent devices, but
are still not intended for use with preexisting devices. The exception
to this is that preexisting container devices can be used as-is.

The error handling has also changed. If an error occurs, the entire
devicetree is restored from a backup, which means that references to
devices in the old tree must be updated in this event.