Wipe partitions before they are created (#950145)
If a partition is created in the same position as a previous filesystem,
eg. a LVM pv, udev will activate it and cause the device to be busy,
preventing us from using it.

This uses the parted information from parted to calculate the start of
the partition and calls dd to write zeros to the first 1MiB of the
partition.

This also rescans the pv to make sure things stay in sync with the
filesystem when creating and removing pvs.