Add classes for lvm thin pool and thin volume.
This is another layer of lvm that provides thinly-provisioned volumes.
A thin pool is a special kind of lv that is a container for thinly-
provisioned lvs. The sizes of the thin lvs are not constrained by the
size of the pool that contains them except at the point where they
actually contain/use that much data/space.