Add support for automatic partitioning using lvm thin provisioning.
The layout starts out the same as the normal lvm autopart layout up to
the vg configuration. A normal swap lv is created the same way as in
normal lvm autopart, and the remainder of the vg is allocated to a
thin pool. From that, we allocate thin lvs for root and possibly home
using the same logic as for normal lvm. So, same as normal lvm autopart
except that root and home are thin lvs.