Update format of iscsi device becoming multipath member (#1039086)
Resolves: rhbz#1039086

We repopulate device tree after add iscsi device dialog in which a login to
some target happened. One issue with this is that format of iscsi device
populated earlier that would make a multipath member together with iscsi device
added in later run of iscsi dialog is not updated to be of multipath_member
type.

So we'd need either 1) require adding of all multipath members in one add iscsi
dialog instance, which would require reworking UI to allow adding nodes of more
targets (ie targets with different IP reachable via different subnets) in one
iscsi dialog; or 2) update the original iscsi device format in devicetree.