Do not try to align partitions to optimal_io_size. (#989333)
Use minimum_io_size instead.

optimal_io_size corresponds to the OPTIMAL TRANSFER LENGTH field in the scsi
spec. It is an upper limit for optimal transfer length and it may be too big to
try to align to it. Use minimum_io_size instead, which corresponds to the
OPTIMAL TRANSFER LENGTH GRANULARITY field in the scsi spec. It is the minimum
size for optimal transfer speed.

Signed-off-by: David Lehman <dlehman@redhat.com>