Convert everything to use Size.
Size essentially stores sizes in bytes and provides methods for
conversion between various units.

All sizes originating from within blivet are based on binary prefixes
except where required to interface with other tools (see ntfsresize).

This adds a missing lower bound of 1 MiB to the md raid superblock size
calculation, which changes some of the mdraid tests.

It also reverses the order of the operands in the various
RAIDLevel._get_raw_array_size methods to ensure we return a Size instead
of an int.