Fix detection of lvm setups. (#1026466)
Fixes two regressions introduced by
commit 8ef070b7b6fe4f270cc0120e5425299abf90b8cd

Because of how space is calculated, setting copies to 0 means the LV
occupies no space in the VG, which is obviously not so.

Also, the spare metadata LV's size is handled as part of the 20% pad
in the VG for pool metadata. Counting it in both places causes more
problems since we didn't take it into account when creating the pool.
(cherry picked from commit 6897f50181662d91d1b6c84e749afe3ec48d2310)