Make get_device_format_class return None if class not found (#1043763)
Related: rhbz#1043763

Previously, get_device_format_class returned DeviceFormat if no precise
class was found for its argument, fmt_type, now it returns None. This brings
its behavior into alignment with the expectation of calling code, all of
which expect that it might return None, except for the one method,
_parseOneLine discussed below.

getFormat() has been changed so that its behavior is exactly the same as
before; it handles the return value from get_device_format_class
differently.

The comments of both functions have been updated.

The only behavioral change is in method _parseOneLine of class FSSet.
Previously, if no format class was found for "nodev", device was set to a
NoDevice object.  Now, the device is not set.  This is in line with the
apparent purpose of the code, which is to set device to a NoDevice object if
fstype indicates a format that is a subtype of NoDevFs.

For reference the if condition that would allow this function to be
unchanged in behavior with the changed behavior of get_device_format_class
is:

if devspec == "none" or fmt_class is None or isinstance(format, fmt_class):

but I think that would not be what was originally intended.

Signed-off-by: mulhern <amulhern@redhat.com>

Conflicts:
	blivet/formats/__init__.py