Set sysfsPath of LUKSDevice when adding to device tree (#1019638)
Resolves: rhbz#1019638

Previously, the LUKS device set up in handleUdevLUKSFormat did not have
its sysfsPath field set, now it does, by calling the updateSysfsPath()
method.

The value of the sysfs_path can not be extracted from the arguments to
handleUdevLUKSFormat, because the info argument is for the parent of this
device, not for the device itself.

Signed-off-by: mulhern <amulhern@redhat.com>