Handle nested btrfs subvolumes correctly. (#1026210)
This is a combination of two commits from master:

Subvolumes are nested. That is to say that removing subvolume foo
will also remove subvolumes foo/bar and foo/bar/baz.
(cherry picked from commit 7a9095de206b38a839e0b38a1dd3d33485253ae3)

Handle passing a btrfs volume as device to BTRFSFactory.

Normally the volume is a container, but it is possible that it has
a mountpoint associated with it, which means it could be passed as
the device in that case.

(cherry picked from commit 21a2e1500c36ff553568d38c9c25dfbfcb33b6df)