Add some basic unit tests for MDRaidArrayDevice.
Add some basic tests to check the state of an MDRaidArrayDevice
after initialization and to check the behavior of some methods.

These tests are added mostly to detect regressions during an anticipated
refactoring.

These tests are very much after the fact and capture the current
behavior, not necessarily the intended or correct behavior.

Therefore, it is fairly likely that if some change to the code
causes a failure, it may be that the test itself is what has
been shown to be wrong.

In some cases, the class was internally inconsistent, in these cases
the inconsistency was resolved. In deciding how to resolve these
inconsistencies more significant methods and simpler code was considered
to have greater credibility than less significant methods or more
complicated code. Errors having to do with None having no attribute
or not being iterable were considered to be unintentional and given
low credibility.

The changes to the tested code are:

* Check whether a smallestMember actually exists and set
smallestMemberSize to 0 if one does not exist or if there are no devices.
Still go through the checks for level even if smallestMemberSize is 0
in order to catch invalid raid level.

* In rawArraySize for RAID10, divide by 2 instead of 2.0. I believe that
if there are three disks available then we get only one disk's worth of data.
Do the same in size.

* A semantics preserving change to initializer.
This change preserves the semantics of the if statement completely.
The advantage is that self.level is no longer
set to any arbitrary value which allows error checking in self.level
setter itself.

* Make level setter error checking.
Previously, the level could be set to anything in the initializer.
Now that this is not true, it is possible to check for invalid values in
the level setter. Valid values are an integer RAID level, "container", or
None.

* smallestMember is altered in a semantically equivalent way, except that
if not self.devices evaluates to True an exception will not be thrown.
Previously, it would have been thrown if self.devices was not a list.
It uses min instead of smallest, since we only need the minimum element
in self.devices.

* totalDevices is also semantically equivalent, but the len of self.parents
is only calculated if needed.

* description is extended so that it makes a somewhat sensible string if
self.level is None.

* In initializer,
use memberDevices setter rather than assigning to _memberDevices
directly in order to get the error checking in the setter.

* formatArgs now returns its formatArgs value