A few small tests for MDFactory class.
These tests are added mostly to detect regressions during an anticipated
refactoring.

These tests are very much after the fact and capture the current
behavior, not necessarily the intended or correct behavior.

Therefore, it is fairly likely that if some change to the code causes
a failure, it may be that the test itself, and the code previously, have
been shown to be wrong.

Signed-off-by: mulhern <amulhern@redhat.com>