Changed devicelibs.mdraid to make use of devicelibs.raid.
Added a Container class to represent mdadmin's "container" level.
Added a Container object to the RAIDLevels object.

Extended raid.RAIDLevels so that it required that all RAIDLevels define
min_members.

Methods changed are:

* raidLevel
* raidLevelString
* get_raid_min_members
* get_raid_max_spares
* get_member_space

All the raid constants now depend on raid.

Methods added are:

* getRaidLevel

* raidLevelString was modified so that it takes a flag to use nicknames
instead of canonical names and so that it accepts any valid descriptor,
not just a numeric one.

Members removed are:

* isRaid
* raid_descriptors

Conflicts:
	blivet/devicelibs/mdraid.py