Change MDRaidArrayDevice to use raid package.
The basic change is that the only call to a function in the mdraid
package is the call to mdraid.getRaidLevel which sets the level field.

After that, all calls are method calls on the level field, so the definitions
of the methods are in the raid package.

Most of the changes did not change the external behavior of the class.

Significant changes are:
* The definition of level was changed.
* rawArraySize now does the same as size and just returns 0 if there
are no member devices.
* Instead of logging an error if the level is container, now rawArraySize
throws an exception, because the definition of the method in the Container
class throws an exception.

Several tests had to be updated, mostly due to the change in the definition
of the level property.