Use given format type as format's name instead of type
Format's type is a strictly set value we define in our classes. Format's name is
a description that can be presented to users instead of type. If we get type
from the outside and cannot find a matching class in our formats, we should set
the type as format's name and let the type unset because many pieces of code
rely on the type being None in case we don't understand/support the format.

Related: rhbz#1033778