BTRFSSubVolume.volume checks the class of its return value.
If it is not an instance of BTRFSVolume a DeviceError is raised.

It must be a BTRFSVolume because that is how btrfs works and also because
otherwise the constructor will raise an AttributeError when calling
_addSubVolume on an object other than a BTRFSVolume.

This change causes the constructor to fail a bit more informatively and a bit
earlier than it otherwise would.

Signed-off-by: mulhern <amulhern@redhat.com>