Show Invalid Disk Label for damaged GPT (#1020974)
A GPT labeled disk could have a damaged partition table, but still be
useful. If that is the case we don't want to treat it like an empty
disk. With these changes it will show up as 'Invalid Disk Label' which
will be translated.

This patch also checks for partitions in the lvm filter that depend on
devices being operated on. If a device is in the action list it
is removed from the lvm reject filter.