Remove redundant checks in existing resize() methods.
It is expected that the resize methods will only be called via the
ActionResizeDevice.execute method. Since this is the case, the removed
checks are redundant.

If it is possible to call the resize methods not through
ActionResizeDevice.execute then the checks that were removed
are not nearly sufficient.

Some comments have been updated, too.

Signed-off-by: mulhern <amulhern@redhat.com>