Remove some methods from devicelibs.btrfs.
* scan_devices method.
It is unused in the codebase.
A sentence on the wiki (https://btrfs.wiki.kernel.org/index.php/Using_Btrfs_with_Multiple_Devices) says:
btrfs device scan is used to scan all of the block devices under /dev and probe for Btrfs volumes. This is required after loading the btrfs module if you're running with more than one device in a filesystem.
* take_snapshot
Totally unused and doesn't do anything.

Signed-off-by: mulhern <amulhern@redhat.com>