Work on devicelibs.btrfs methods that require that the device be mounted.
Make a small subclass that handles mounting the device before tests
and unmounting after tests. I pondered making this a more general class
that can be shared among the different filesystem tests, but there are
some things about mounting the filesystem that are somewhat btrfs specific.

Otherwise just tests and comments.

subvolume list is pretty configurable and its output can change somewhat
depending on the version installed. The tests that use it are written
to be robust to small changes, but could probably be strengthened a bit
as btrfs settles down.

Signed-off-by: mulhern <amulhern@redhat.com>