Changes to the handling of filesystem labeling (#1038590)
Resolves: rhbz#1038590

* Change FS.label to a property
- Want to guard setting the label format, so it matches the requirements of
the labelfs program that will use it.
- Don't set label property in writeLabel. writeLabel is only called from
doFormat, in a context where self.label is guaranteed to be set.
* Removed _defaultLabelOptions field and defaultLabelOptions property. The
property is only used in this file, and it is only used by _getLabelArgs.
The property is set from the field.
* Remove label parameter from writeLabel method
- writeLabel is only called in one place and in that place self.label is
guaranteed to be set, so this change does not change the behavior of the
program.
- Since writeLabel uses the label in self.label instead of a parameter
it is guaranteed that this value has passed the checks in self._setLabel.
- Other methods that are only called by writeLabel are changed in the same way.
- writeLabel now throws an exception if it can not set the label because there
is no program to set it. This does not change the behavior of this program.
* Factor out filesystem relabeling functionality into a separate package
* Add the ability to read the filesystem label
- In some cases, the application that is used to write the label is also
used to read the label.
- This is certainly useful for testing, and may come in handy otherwise.

Signed-off-by: mulhern <amulhern@redhat.com>

Conflicts:
	blivet/formats/fs.py