Make a class that creates a unique-per-class id for objects (#1043763)
Related: rhbz#1043763

There are some objects that we would prefer to be able to track across a copy
or a deep copy. This is most likely when numerous instances of
more or less indistinguishable objects are created and when they are likely to
be copied during an anaconda run. DeviceFormat objects are the motivating
example.

Signed-off-by: mulhern <amulhern@redhat.com>