Add a labelFormatOK method to the DeviceFormat's interface (#1038590)
Related: rhbz#1038590

Need to expose this information to any client of a DeviceFormat since it's
really an operational question about the labeling application not about
the state the DeviceFormat is allowed to have.

In other words, have to allow for a legal DeviceFormat label which the labeling
application can not handle.

Signed-off-by: mulhern <amulhern@redhat.com>