Move label setter and getter into DeviceFormat class (#1038590)
Related: rhbz#1038590

DeviceFormat does not have a label but SwapSpace as well as FS does, and
it is better to put the same restrictions on the label for both.
Note that the restriction has been reduced to converting "" to None.
The comments on the setter explain why it is not possible to be more
restrictive.

Signed-off-by: mulhern <amulhern@redhat.com>