Allow resetting partition size to current on-disk size. (#1040352)
The test to see if the new size is different from the currently
projected size was erroneously using the currentSize attribute instead
of the size attribute. The currentSize attribute is the wrong choice
because it always reflects the device's actual on-disk size, meaning
it ignores any previously scheduled resize actions and makes it
impossible to reset the device's targetSize back to its actual on-disk
size once it has been set to something else.