Remove BTRFS._getFormatOptions (#1038590)
Related: rhbz#1038590

It can not be called when the the instance is BTRFS because only
FS.doFormat calls it and doFormat is only called by create which fs.BTRFS
does not use.
So, it can not be called at all.

Also, BTRFS is the only class that overrides this method, so it's existance
is a bit distracting.

Signed-off-by: mulhern <amulhern@redhat.com>