Change the meaning of label field values (#1038590)
Related: rhbz#1038590

Now, if FS.label is None that means accept the filesystem default and if it
is "" that means write an empty label.

If a filesystem exists when the object is created, the label should not be
None, as that filesystem must have some label even if it is the empty string.

Signed-off-by: mulhern <amulhern@redhat.com>