Fall back on relabeling app where available (#1038590)
Related: rhbz#1038590

If there is a relabeling app, don't use filesystem creation app to label
the filesystem, use the relabeling app.

The motivation here is that it is better for labeling to fail separately
when possible than for filesystem creation to fail due to issues with
labeling.

The call to writeLabel is inside the if because I'ld rather not log
a failure due writeLabel() failing to write the label None, which means
accept the default and do not write a label. I'ld also rather not log
a failure when the label has already been written by the creation app
because there is no separate labeling app.

Signed-off-by: mulhern <amulhern@redhat.com>