Look up SELinux context for lost+found where it is needed (#1038146)
Resolves: rhbz#1038146

Previously the context for lost and found was set at module import as a module
level file, but it was only used when mounting a filesystem. Now, it is set
where it is used and the lookup uses a little method in util.py instead of the
raw selinux method.

Signed-off-by: mulhern <amulhern@redhat.com>