When repopulating multipath members mark them as multipath (#1056024)
Resolves: rhbz#1056024

This is needed when doing (re)populate when some FCoE multipath devices had
already been added in previous populate. We need to add info["ID_FS_TYPE"] (as
when the device is scanned/populated for the first time) so that the
condition excluding multipath menbers in handleUdevDiskLabelFormat is
applied to the device.