Tell lvcreate not to ask us any questions and do its job. (#1057066)
Specifically, new functionality prompts the user whether or not they
would like to wipe detected metadata when creating a new lv. This
tells it to go ahead instead of asking.

This is a global setting for lvcreate, but I think we really do want
to create the new lv in any case, regardless of what lvm might think
we should stop and talk about first.