provide a more useful error message if user fails to create an ESP
If you do a UEFI native installation, use custom partitioning, and fail to
mount an EFI system partition at /boot/efi, blivet's __init__.py attempt to
set a stage1 device will fail completely and return our old friend, the
completely oblique error message "you have not created a bootloader stage1
target device". With this patch, when that test fails, if the arch is UEFI,
an additional error message is appended: "For a UEFI installation, your layout
must include an EFI system partition mounted at /boot/efi". It's always hard
to consider all possible paths to an error condition, but I don't _think_ it
would ever be incorrect to display this message when the attempt to set a
stage1 device fails during a UEFI installation, and usually it will be a
useful cue for the user.

Signed-off-by: Will Woods <wwoods@redhat.com>