escrow: make sure the output directory exists (#1026653)
writeEscrowPackets() gets called from turnOnFilesystems() immediately
after it mounts all the filesystems. So under normal circumstances, when
we run writeEscrowPackets() the ROOT_PATH is completely empty.

So when we do device.format.escrow(ROOT_PATH + "/root",...) that path
doesn't exist yet, and so it ends up raising IOError when it tries to
open a file there.

The simple solution is to ensure the path exists before we pass it to
device.format.escrow().

(This won't interfere with the installer; when RPM tries to create
/root in the course of installing the "filesystem" package it should
happily just chmod the existing directory to the proper mode and leave
the contents as-is.)