Provide a mechanism for platform-specific error messages for stage1 failure
This is a follow-up to "provide a more useful error message if user fails to
create an ESP". It implements a cleaner, more generic way of doing it: instead
of stuffing a check for UEFI platform and a platform-specific error message
into the storage sanity check, it provides a mechanism by which we can specify
in platform.py an error message to be displayed in the case the stage1 check
fails.

I've included error messages for each platform, based on the constraints.

I also slightly tweaked the text of the generic error message itself to be
more accurate to the logic as it currently stands: stage1_device not being set
doesn't necessarily mean "nothing and no-one tried to set it" any more, it can
also mean "we tried to set it but couldn't find a valid setting" - in fact,
that's much more likely. "See below for details" makes it a bit clearer, I
think, that the platform-specific error message is giving you more details
about this still-fairly-cryptic message, they're not separate errors. "stage1"
is pretty jargon-y, probably meaningless to the user, and probably not strictly
the right term for some of these cases, so I dropped it.