Make a small SanityException hierarchy (#1060255)
Related: rhbz#1060255

Class hierarchy looks like this:

StorageError
   SanityException
      SanityError
         LUKSDeviceWithoutKeyError
         ... more stuff here, as needed ...
      SanityWarning
         ... more stuff here, as needed ...

SanityException should be something generated by blivet.sanityCheck.
LUKSDeviceWithoutKeyError is currently the only distinguished subclass of
SanityError. It's special,
because in some situations where sanityCheck is called, a missing key
should not be considered fatal, but it's the caller's job to figure out if
that is the situation.

Signed-off-by: mulhern <amulhern@redhat.com>