Change messages to SanityExceptions objects (#1060255)
Related: rhbz#1060255

Change string messages returned in sanityCheck tuple and in
_verifyLUKSDevicesHaveKey to Error objects.
The major purpose of this is to be able to distinguish and filter them.

Use object type to distinguish warning (SanityWarning) from errors
(SanityError) instead of passing back a pair of lists.

Also, got rid of that large chunk of commented out code, if we want it, we
know where to find it.

Signed-off-by: mulhern <amulhern@redhat.com>