Improve handling of parted.DiskLabelCommitError slightly.
This approach is more focused in that it only deactivates devices that
actually use the disk we're trying to commit to.

This brings up something that should probably be addressed at some point:
We should consider some kind of warning when a partition table change is
scheduled on a disk with in-use partitions.