Fix partition handling across multiple processActions calls. (#1065522)
We replace the parted.Disk instances in all DiskLabel instances every
time processActions is run, so we need update the parted.Partition
instances throughout the devicetree accordingly.

Also, we have to make sure that each DiskLabel's _origPartedDisk is
updated before we update the DiskDevice's originalFormat with a new
copy of the DiskLabel.