Count with the extra metadata extents for RAID consistently (#1065737)
When creating an LVM setup on top of RAID, we add some extra extents to make
sure the metadata will fit in. However, when reporting free space in a VG or
when creating an LV in a VG the size of which was pushed to the limit, we need
to take those extra metadata into account as well.