Fix handling of extended partitions across various modes of operation.
When flags.installer_mode is True we only create extended partitions as
needed and we remove them each time we allocate partitions. When
flags.installer_mode is False, extended partitions can be created
explicitly, and can only be removed as part of partition allocation if
they were not explicitly requested.