Make DeviceTree.hide() remove a larger set (#1043763)
Related: rhbz#1043763

If a device is hidden these things now happen:
1) All actions are canceled. Note that canceling a device create action will
cause the device to be removed. Note that it should be the case that no
existing device has a device create action associated with it.
Previously only a subset of actions were canceled.
2) The device being hidden is removed using _removeDevice. An additional
post-processing step adds the name of the device back into self.names.
Other steps associated with hiding the device remain the same.

Note: The device exists if and only if it has not been removed by the
cancellation of all actions. So, if it does not exist, then the method
is done after action cancelation. However, if the device does exist, it should
be removed, special hiding actions should be done, and its name should be
inserted back into self.names.

Signed-off-by: mulhern <amulhern@redhat.com>