Make crypt tests runnable.
If any of these tests seem wrong then the code can be fixed up and the test
changed.

A tearDown method was added to close the device in case a test failed.

Some instance fields were used for the names of the devices.

Some aspects of the code that brought about these changes are:
* All crypto.luks_* methods raise a ValueError if they are not given a
passphrase although they accept a key_file argument.
* If a device does not exist, CryptSetup raises an IOError.

Some methods that did suprising things were.
* is_luks --- it raises an IOError for an unknown device
* luks_format --- raises a ValueError if it gets a keyfile instead of a
passphrase
* luks_remove_key --- raises a CryptoError instead of a RuntimeError
* luks_status --- returns ints, not booleans

Signed-off-by: mulhern <amulhern@redhat.com>