Extend the list of things to be omitted if moddisk is False (#1043763)
Related: rhbz#1043763

If moddisk is False then avoid removing partitions, but also avoid removing
logical volumes and subvolumes.

The idea is to leave information about what is on the device intact, but to
remove it from the list of devices.

Only the top few lines have actually changed, the rest is indentation only.

Signed-off-by: mulhern <amulhern@redhat.com>