Only log a warning about labeling if something is wrong (#1075136)
Resolves: rhbz#1075136

If do_labeling is False, then we are not choosing to set a label during
filesystem creation. Probably we are deferring setting a label until
after the filesystem has been created.
If self.label is None then we are accepting the filesystem default label,
and not setting any label.

So, we only log a warning if setting a label is what ought to happen but we
don't want to risk it because it looks like the label format is not accepted
by this filesystem and that therefore, specifying the label arguments may cause
failure at filesystem creation.

Note: Some filesystem creation apps will truncate a label that is just too
long but others will fail instead. We do not want to distinguish
among filesystems based on this behavior.

Signed-off-by: mulhern <amulhern@redhat.com>