Allow NTFS to be mountable. (#748780)
Back in the time of Fedora 14, anaconda marked NTFS volumes as
unmountable, because any attempt to mount them would hang in the mount
wrapper in imount.c. Because NTFS is handled through fuse, an instance
of mount.ntfs stays running after the mount operation is completed, and
mount.ntfs does not property daemonize: it leaves stdout and stderr
open, so we were getting stuck trying to read from what should have been
closed file descriptors.

mount.ntfs still does this, but the problem has been twice fixed
elsewhere: /bin/mount closes its stdout and stderr--which closes the
copies attached to anaconda's pipes--when it exits, and the python code
we now use in place of imount pays attention to SIGCHLD, so it stops
trying to read output when mount exits whether the file descriptors are
still open or not. So we can mount ntfs partitions now if anyone wants
them.