Fix logging function string format warnings.
Specify parameters to logging function strings as parameters to the
logging function so that the format operation does not need to be
evaluated for message levels that will not be logged.

Signed-off-by: mulhern <amulhern@redhat.com>

Note that this can mean the difference between crashing and not-crashing,
if your __repr__ function is unkind enough to throw an AttributeError,
for instance.