Suppress some unused variable warnings
Suppose we have some code:

for (x, y, z) in <some list of tuples>:
    # do something with x
    # do something with y

Then z will be reported as an unused variable. But really, there is
no more elegant way to give x and y names, so it's better to supress
the warning about z.

We do that by preceding the variable name with an underscore,
like _z.

Signed-off-by: mulhern <amulhern@redhat.com>