Remove unused assignment to boot.
boot is not used anywhere in the method below and I do not believe that it
is global. So, the only effect of this change would be that if
self.bootDevice raises an exception that is not in
(DeviceError, AttributeError) that exception will no longer be propagated.
But it seems unlikely that such an exception was being raised anyway.

Signed-off-by: mulhern <amulhern@redhat.com>