Suppress some unused variable warnings.
Suppress warnings when the unused variable is in a tuple that matches
the result of a split() or partition() operation and the possibility
of a failure to match is handled appropriately.

Note that sometimes a failure to match is fine and does not explicitly
need to be checked for as in matching an optional comment.

Suppress by prepending unused variables with an underscore.

Signed-off-by: mulhern <amulhern@redhat.com>