Split ROOT_PATH usage into getTargetPhysicalRoot()/getSysroot()
This mirrors the change to Anaconda, in order to enable OSTree
support.

For OSTree, the location of the OS checkout (and e.g. /etc/fstab) is
really in /ostree/deploy/$osname/deploy/$revision/etc/fstab.

In order to properly support OSTree, Blivet will need to gain an
understanding of the separation between the physical system / and the
target root.

This patch will be used in Anaconda, which will call setSysroot()
attribute after the root being installed is laid out.

After that, when we call write(), the fstab data will be correctly
written into the target root.