Force removal of hidden devices (#1078163)
Resolves: rhbz#1078163

An extended partition is considered a non-leaf device if the parted info
indicates that it contains some logical partitions.

In hide, we need to preserve the parted info, when we are removing existing
devices, in order to reconstruct it later if we unhide the device.

Therefore, in hide, when removing a PartitionDevice, we are not changing the
parted information, so we must force the removal to avoid throwing the
non-leaf exception.

In theory, we only need to force the removal for PartitionDevices that happen
to be extended, but making the force conditional in that way is probably
excessive.

Signed-off-by: mulhern <amulhern@redhat.com>