Change management of Device parents to use a simple list interface.
The parent list is an instance of class ParentList, which provides a
simplified list interface with the addition of an element uniqueness
constraint and optional user-defined pre-add and pre-remove functions.

Before, to modify a device's parent set, you'd do this:

 device.parents.append(newparent)
 newparent.addChild()

or this, if device is a container type:

 device._addMember(newparent)

Now, regardless of the type of device, you do this:

 device.parents.append(newparent)

Any checking or accounting is handled by the callbacks registered to
the parent list.