Improve the mechanism for VG completeness.
Do not modify self.pvCount -- that makes it meaningless as it was
obtained by lvm.

Do not worry about completeness if no pvCount was passed in for an
existing VG.

Because we instantiate all Device classes with exists=False when unit
testing, add a hack to StorageTestCase.newDevice to set completeness
appropriately for existing VGs.