Exit if the specified log file does not exist.
If the log file does not exist, diff-quality will try to make it by calling
pylint. diff-quality passes some arguments to pylint which some versions
of pylint fail to recognize. In the event that one of these versions of
pylint is running diff-quality will raise an exception.

Better to catch it in the script than to leave diff-quality to its own
devices.

Signed-off-by: mulhern <amulhern@redhat.com>