Suppress false positive W0631 error.
The only way the else branch is not executed is if the break statement is
encountered. Since the else branch exits the only way this line can be
reached is if the break statement is encountered. But for the break
statement to be encountered, enumerate lines must be non-empty, and
therefore i,l must have been assigned to when the line is reached.

Not putting in pylint-false-positives because this may prove to be beyond
the scope of the W0631 analysis.

Signed-off-by: mulhern <amulhern@redhat.com>