Fix a bug and catch a change in lvm's thin pool layout.
The first chunk fixes what appears to be something new in thinp --
using eg: vg-thinpool_tmeta instead of vg-thinpool_rmeta%d as the
metadata volume name.

The second chunk drops the addition of the vg name to keys in lv_info,
which already contain the vg name, when constructing the raid dict.