Fix accounting related to addition of md member devices.
We should not be incrementing memberDevices if we're just adding a member
device as part of devicetree.populate. In that case we already know the
number of memberDevices and shouldn't be changing it. It should be
incremented when adding a new member to an existing array, like when
scheduling a ContainerAddMember action.