Suppress W0201 errors
The message for W0201 is
"Attribute '<attribute>' defined outside __init__".

In these cases, the attributes are certainly defined during
execution of __init__, but through calls to other methods, i.e., not
directly in __init__.

Not putting this in false positives because I doubt that pylint will be able
to develop the necessary call graphs to do the analysis.

Signed-off-by: mulhern <amulhern@redhat.com>