Round sizes in humanReadable instead of flooring them
humanReadable representation of a Size instance can round the values instead of
flooring them because it should never be passed to any code requiring the
precise size. And it's much nicer to give user '9.98 GiB' instead of '9.97 GiB'
when they enter '9.98 GiB'

Also get rid of the unnecessary _trimEnd() method, the Decimal() constructor can
do the same for us.