Change signature of DiskLabel.addPartition to be more useful.
The new method accepts start sector, end sector, and partition type.

The previous method required you to pass in a parted.Partition instance.
It did not simply add that parted.Partition instance. Instead, it created
a new parted.Partition instance based on the one passed in and added that
to the parted.Disk.