Update Platform instance from flags instead of replacing it. (#1090646)
Replacing the platform instance as a means of picking up changes to
flags doesn't work because modules that have already explicitly imported
platform from platform will still have the now-defunct Platform instance
in their namespace. This patch changes the flags update so that it
updates the Platform instance in-place, allowing the modules that have
already imported it to get the updates.