Disable unused argument warning for 'disks' in get_pv_space.
The original version of this method had a mandatory argument, 'disks', which
it did not use. All methods that invoke get_pv_space pass an argument for
disks. I assume that 'disks' is relevant to the result value but the formula
in the method has not become sophisticated enough to incorporate it.

Signed-off-by: mulhern <amulhern@redhat.com>