Disable a no member warning for EddTestFS initializer.
This warning arises because EddTestFS depends on a mock package in the
anaconda/old_tests directory. The majority of the tests in edd_test.fs would
fail if run due to this dependency, but they are all being skipped.
It is possible that they could be fixed by grabbing the mock package or by
some other means.

Signed-off-by: mulhern <amulhern@redhat.com>