Fix handling of devices activated as a side-effect of actions.
Removing a partition triggers a udev change event for the disk, which
then triggers udev rules which activate all md arrays on that disk.

getDependentDevices doesn't always work because of various effects of
action registration. Perhaps most importantly, devices that exist but
whose removal is pending are not in the list of devices checked by
getDependentDevices.